backend/pyproject.toml
[project]
name = "yt-mcp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "jupyterlab>=4.4.9",
    "mcp[cli]>=1.15.0",
    "openai>=1.109.1",
    "openai-agents>=0.3.3",
    "python-dotenv>=1.1.1",
    "praw>=7.7.1",
    "requests>=2.32.0",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "pydantic>=2",
]
